requirements.txt
simpy
pandas
jupyter
duckdb
streamlit
kafka-python
pyspark